refactor(horus): split the facade by module, and drop what nothing reaches
horus/src was two files holding five public modules, and lib.rs was 868
lines of which 268 were code. Each public module now has a file named
after it, so the path a reader wants is the file they open:

  lib.rs      499  crate landing page + the crate-root re-exports
  prelude.rs   98  the one glob import
  msg.rs      113  the horus::msg:: namespace C++ and Python also write
  types.rs     39  msg plus the machinery under the payloads
  net.rs      121  the optional LAN replication, was inline behind a cfg
  time.rs     384  unchanged

No public path moves, so this is not a breaking change: `horus::msg`,
`horus::types`, `horus::prelude` and `horus::net` resolve exactly as
before. Moving a module's docs from `///` above the declaration to `//!`
inside the file does re-scope its intra-doc links, which rustdoc caught
under -D warnings; those are now `crate::`-qualified.

Deleted, after checking every consumer:
  assets/logo.png     1.36 MB, orphaned since 40cca875 (2026-03-30)
                      removed the README <img> that referenced it. No
                      .md, workflow, Cargo metadata or packaging names
                      it, and the package root is horus/, so it was in
                      neither the .crate nor the wheel.
  qa_crash_safety.rs  no referent anywhere in the tree
  qa_blackbox_record.rs   likewise
  [lib] in Cargo.toml restated cargo's own defaults

Kept deliberately: estop_trigger.rs and launch_ctl_send.rs are also
unreferenced, but they are the only in-repo exercisers of the launch
e-stop and launch-control topics. Deleting them would remove the last
thing that touches those paths; they need a test, not a delete.

Fixed while here:
  README.md      message!(SensorData = (f64, u32)) cannot compile — the
                 macro has no tuple arm and its own error text says so
  three examples reached past the facade with horus_core:: paths that
                 do not resolve in a real user project, where the
                 prelude already works
  lib.rs         'Common Mistakes' #4 said read_latest() needs T: Copy
                 'for multi-consumer backends'. Both impls require it
                 unconditionally (topic/mod.rs:3427, :4444)
  lib.rs         [`scheduling::NodeBuilder`] rendered as literal text —
                 scheduling is #[doc(hidden)], so the page never existed
  Cargo.toml     ctor was a non-optional dependency the default build
                 never uses; now gated behind `net`, the only consumer.
                 serde moved to dev-dependencies: the library does not
                 link it, the examples and tests do.

Verified: fmt, clippy -D warnings (workspace), cargo doc -D warnings,
cargo check --workspace --all-targets, cargo check -p horus with
--no-default-features and with --features net, and all 8 horus test
suites (76 tests).
