+++ horus/src/lib.rs
@@ -144,7 +144,8 @@
 //! | `.async_io()` | **AsyncIo** | Network calls, file I/O, logging (tokio pool, no RT impact) |
 //! | *(default)* | **BestEffort** | Diagnostics, telemetry (main thread, sequential) |
 //!
-//! See [`scheduling::NodeBuilder`] for the full decision guide.
+//! The builder returned by `Scheduler::add` carries the full set; `.build()` is
+//! what registers the node once you have chosen one.
 //!
 //! ### Key Message Types
 //!
@@ -187,7 +188,9 @@
 //! `.build()?` at the end silently drops the node registration. Always end with `.build()?`.
 //!
 //! **4. `read_latest()` on non-Copy types** — `Topic::read_latest()` requires `T: Copy`
-//! for multi-consumer backends. Use `recv()` for non-Copy types like `String` or `Vec`.
+//! on every backend, not just the multi-consumer ones: it hands back a copy of bytes
+//! another process may overwrite at any moment, so a `T` owning a heap pointer would
+//! be a use-after-free. Use `recv()` for non-Copy types like `String` or `Vec`.
 //!
 //! **5. `.rate()` order doesn't matter** — `.rate(100.hz()).compute()` produces Compute,
 //! not Rt. Execution class is determined by the explicit class call, not by `.rate()`.
@@ -225,129 +228,7 @@ fn __register_network_auto_wire() {
 }
 
 #[cfg(feature = "net")]
-pub mod net {
-    //! Transparent LAN replication — zero config, same `Topic<T>` API.
-    //!
-    //! Network replication starts **automatically** when `scheduler.run()` is called.
-    //! No manual wiring needed:
-    //!
-    //! ```rust,ignore
-    //! use horus::prelude::*;
-    //!
-    //! let mut scheduler = Scheduler::new()
-    //!     .tick_rate(100_u64.hz());
-    //! // ... add nodes ...
-    //! scheduler.run()?;
-    //! // Network replication starts automatically and stops on shutdown
-    //! ```
-    //!
-    //! ## Opting out
-    //!
-    //! Disable networking via `.network(false)` or `HORUS_NET_ENABLED=false` env var:
-    //!
-    //! ```rust,ignore
-    //! let mut scheduler = Scheduler::new()
-    //!     .network(false)              // <-- disables LAN replication
-    //!     .tick_rate(100_u64.hz());
-    //! ```
-    //!
-    //! ## Custom config
-    //!
-    //! For advanced networking configuration, use [`wire_with_config()`] to override
-    //! the automatic default:
-    //!
-    //! ```rust,ignore
-    //! use horus::net::NetConfig;
-    //!
-    //! let config = NetConfig { port: 9200, ..NetConfig::default() };
-    //! horus::net::wire_with_config(&mut scheduler, config);
-    //! scheduler.run()?;
-    //! ```
-    //!
-    //! ## Manual control
-    //!
-    //! Call [`enable()`] before `scheduler.run()` and hold the returned handle:
-    //!
-    //! ```rust,ignore
-    //! let _net = horus::net::enable();
-    //! scheduler.run()?;
-    //! ```
-
-    /// Enable network replication with default settings.
-    ///
-    /// Returns a handle that stops the replicator on drop.
-    /// Returns `None` if networking is disabled (`HORUS_NO_NETWORK=1`).
-    pub fn enable() -> Option<horus_net::ReplicatorHandle> {
-        horus_net::start_replicator_default()
-    }
-
-    /// Enable network replication with custom config.
-    pub fn enable_with_config(
-        config: horus_net::config::NetConfig,
-    ) -> Option<horus_net::ReplicatorHandle> {
-        horus_net::start_replicator(config)
-    }
-
-    /// Wire `horus_net` into the scheduler's lifecycle (manual override).
-    ///
-    /// **Note**: As of HORUS 0.2, networking is wired automatically when
-    /// `scheduler.run()` is called — you do NOT need to call this function
-    /// for default configuration. It remains available for backward
-    /// compatibility and as an explicit opt-in if you want to be explicit.
-    ///
-    /// If networking was explicitly disabled (`.network(false)` or
-    /// `HORUS_NET_ENABLED=false`), this is a no-op.
-    ///
-    /// # Example
-    /// ```rust,ignore
-    /// use horus::prelude::*;
-    ///
-    /// let mut scheduler = Scheduler::new()
-    ///     .tick_rate(100_u64.hz());
-    /// scheduler.add(my_node).build()?;
-    ///
-    /// // Optional: explicit wire (happens automatically if omitted)
-    /// horus::net::wire(&mut scheduler);
-    /// scheduler.run()?;
-    /// ```
-    pub fn wire(scheduler: &mut horus_core::Scheduler) {
-        if !scheduler.network_enabled() {
-            return;
-        }
-
-        scheduler.on_start(|| match horus_net::start_replicator_default() {
-            Some(handle) => {
-                eprintln!("[horus_net] Network replication started");
-                Some(Box::new(handle))
-            }
-            None => None,
-        });
-    }
-
-    /// Wire `horus_net` with custom configuration into the scheduler's lifecycle.
-    ///
-    /// Same as [`wire()`] but allows passing a custom [`NetConfig`].
-    /// When called, the custom config takes precedence over the automatic
-    /// default wiring.
-    pub fn wire_with_config(
-        scheduler: &mut horus_core::Scheduler,
-        config: horus_net::config::NetConfig,
-    ) {
-        if !scheduler.network_enabled() {
-            return;
-        }
-
-        scheduler.on_start(move || match horus_net::start_replicator(config) {
-            Some(handle) => {
-                eprintln!("[horus_net] Network replication started (custom config)");
-                Some(Box::new(handle))
-            }
-            None => None,
-        });
-    }
-
-    pub use horus_net::config::NetConfig;
-}
+pub mod net;
 
 // === Internal plumbing (hidden from docs, used by horus_py / macro-generated code / horus_manager) ===
 #[doc(hidden)]
@@ -414,261 +295,11 @@ pub use horus_core::horus_internal;
 #[doc(hidden)]
 pub use horus_core::serde_yaml;
 
-/// Message and geometry types.
-///
-/// This used to be a plain re-export of the `horus_types` crate, which left
-/// half the types unreachable from the obvious path. `Tensor`, the point-cloud
-/// extension traits and their neighbours live in `horus_core::types`, so
-///
-/// ```rust,ignore
-/// use horus::types::Tensor;              // error: no `Tensor` in the root
-/// use horus::horus_core::types::Tensor;  // what the docs had to say instead
-/// ```
-///
-/// Reaching through `horus::horus_core::` defeats the point of a facade: it
-/// names an internal crate the user was never meant to know about, and it is
-/// not guessable — you find it by reading the source or by copying a doc
-/// snippet. Both sets are re-exported here, so the obvious path is the one that
-/// works.
-///
-/// The same argument then stopped one level short. This module is called
-/// "types", its own heading says "Message and geometry types", and it is the
-/// path a reader guesses before they have heard of [`msg`] — but it held only
-/// the geometry half. `horus::types::Pose2D` resolved while
-/// `horus::types::CmdVel`, `Imu`, `Odometry`, `LaserScan` and `NavGoal` all
-/// failed with `E0425` and no hint about where to look instead. A path that
-/// works for half the names you try is worse than one that does not exist at
-/// all: it reads as confirmation that you are in the right module and the type
-/// is simply missing. Everything in [`msg`] is re-exported here too, so the
-/// guess lands. [`msg`] stays the spelling the documentation uses, because it
-/// is the one C++ and Python already write.
-pub mod types {
-    pub use horus_core::types::*;
-    pub use horus_types::*;
+pub mod types;
 
-    // Glob-re-exporting [`msg`] rather than naming the messages one by one is
-    // what keeps the two from drifting: whatever becomes reachable as
-    // `horus::msg::X` is reachable as `horus::types::X` by construction, so
-    // this module cannot silently fall back to holding half the set. Where the
-    // three globs overlap (`Pose2D`, `ImageEncoding`, `Quaternion`, …) they
-    // resolve to the same item, so no name becomes ambiguous.
-    pub use crate::msg::*;
-}
+pub mod msg;
 
-/// Standard message types, under one path — the same `horus::msg::` spelling
-/// C++ uses.
-///
-/// ```rust
-/// use horus::msg::{CmdVel, Imu, Pose2D};
-/// ```
-///
-/// The cross-language compatibility table used to give the same seven wire
-/// types three unrelated Rust paths — `horus_robotics::CmdVel`,
-/// `horus_robotics::messages::sensor::Imu`, `horus_types::Pose2D` — while C++
-/// wrote `horus::msg::X` for all of them and Python wrote `horus.X`. Which
-/// crate a message happens to live in is a packaging decision (`horus_robotics`
-/// and `horus_tf` are separate git repositories), and it was leaking into every
-/// user's import lines: moving between the three languages meant re-learning
-/// the location of every type.
-///
-/// [`prelude`] already pulls them all in, but a glob is not a path. It cannot
-/// be written in a type position, and it tells a reader nothing about where
-/// `Imu` came from. [`types`] was the near miss that made the guess worse: it
-/// existed, it sounded canonical, and it held only the geometry half, so
-/// `horus::types::CmdVel` failed with `E0425` instead of pointing anywhere
-/// useful. That module now mirrors this one, so neither guess dead-ends.
-///
-/// Every name C++ declares across `<horus/msg/*.hpp>` resolves here — the
-/// `msg_covers_the_whole_cpp_namespace` test below is that header list, so the
-/// two namespaces cannot drift apart silently. The pool-backed payloads and
-/// the transform POD are here too even though C++ files them one namespace up,
-/// in plain `horus::`: `Image`, `PointCloud`, `DepthImage` and `Transform` are
-/// in this crate's own "Key Message Types" table, so a reader who just watched
-/// `horus::msg::CmdVel` work will type `horus::msg::Image` next, and getting
-/// `E0425` there would rebuild the very trap [`types`] used to be.
-///
-/// What is deliberately *not* here is the machinery underneath those payloads:
-/// `Tensor`, `TensorDtype`, `Device` and the raw `*Descriptor` structs. They
-/// are not things you put in a `Topic`, they are what the pool types are built
-/// out of, and they live in [`types`] — which re-exports this module in full,
-/// so `horus::types::` resolves for everything below plus those. Saying so out
-/// loud is the point: an exclusion a reader only discovers as an `E0425` is the
-/// same trap this module was written to remove.
-///
-/// Nothing moved: these are re-exports, so every path that worked before still
-/// works. This only makes one of them canonical, and picks the one the other
-/// bindings already use.
-///
-/// This module is glob-safe next to the prelude — you can write both, and the
-/// one name where the two source crates disagree resolves the same way under
-/// either:
-///
-/// ```rust
-/// use horus::prelude::*;
-/// use horus::msg::*;
-///
-/// // The action-server lifecycle enum, the meaning `horus::prelude` has
-/// // always given the bare name.
-/// let lifecycle: GoalStatus = GoalStatus::Succeeded;
-/// assert!(lifecycle.is_terminal());
-///
-/// // The navigation status code stored in `GoalResult::status` keeps the
-/// // module path the docs already teach for it.
-/// assert_eq!(navigation::GoalStatus::Succeeded as u8, 2);
-/// ```
-pub mod msg {
-    /// Robotics wire types — sensor, control, navigation, detection, vision,
-    /// force, input, perception and simulation. Rust ships a few that have no
-    /// C++ counterpart yet — the simulation service messages, and the vision
-    /// types `<horus/msg/vision.hpp>` deliberately declines to publish.
-    pub use horus_robotics::messages::*;
-
-    /// Math, diagnostics, time and generic messages. This is `horus_types`'
-    /// curated prelude rather than its crate root, because the root also
-    /// re-exports the crate's own module names (`math`, `time`, …) and those
-    /// are an implementation detail of where the types are filed, which is
-    /// exactly what this module exists to stop leaking.
-    pub use horus_types::prelude::*;
-
-    /// The zero-copy payloads. C++ declares these in `horus::` rather than
-    /// `horus::msg::` (`<horus/pool.hpp>`), and `DepthImage` has no C++ type at
-    /// all, so parity offers no answer here — but they are messages in every
-    /// sense that matters to a caller: they are what a `Topic<Image>` carries.
-    /// Leaving them out made `horus::msg::Image` an `E0425` sitting two lines
-    /// below a working `horus::msg::LaserScan`.
-    pub use horus_core::memory::{DepthImage, Image, PointCloud};
-
-    /// The element types you need to fill the payloads above —
-    /// `Image::new(w, h, ImageEncoding::Rgb8)`, `PointCloud::from_xyz(&[PointXYZ])`
-    /// — for the same reason: a namespace that hands you `PointCloud` but not
-    /// `PointXYZ` has only moved the missing import one line down.
-    pub use horus_core::types::{ImageEncoding, PointXYZ, PointXYZI, PointXYZRGB};
-
-    /// The rigid transform. C++ writes `horus::Transform` (`<horus/transform.hpp>`),
-    /// one namespace up, and Rust used to require a hand-added `horus_tf` git
-    /// dependency to name it at all.
-    pub use horus_tf::Transform;
-
-    /// `GoalStatus` is the one name the two source crates disagree about.
-    /// `horus_core::actions::GoalStatus` is the action-server lifecycle enum
-    /// (`Pending`/`Active`/`Succeeded`/`Aborted`/`Canceled`/`Preempted`/`Rejected`);
-    /// `horus_robotics::messages::navigation::GoalStatus` is the `#[repr(u8)]`
-    /// code stored in `GoalResult::status`, with different variants. C++ never
-    /// had to choose — `<horus/msg/navigation.hpp>` publishes `GOAL_STATUS_*`
-    /// constants and puts the enum class in `horus::`, not `horus::msg::` — so
-    /// there is no parity answer to copy.
-    ///
-    /// Left to the glob above, this one name made `use horus::prelude::*;`
-    /// together with `use horus::msg::*;` a hard `E0659` ambiguity error
-    /// (`ambiguous_glob_imports` is deny-by-default), which is the one thing a
-    /// convenience namespace must not do to the prelude it sits beside. Pinning
-    /// it to the prelude's meaning is also the rule the documentation already
-    /// teaches: the bare name is the action enum, and the navigation code is
-    /// reached through its module — `horus::msg::navigation::GoalStatus`, the
-    /// exact shape of the `horus::prelude::navigation::GoalStatus` the
-    /// navigation reference has always told readers to write.
-    pub use horus_core::actions::GoalStatus;
-}
-
-/// The HORUS prelude — everything you need for building robotics applications.
-///
-/// ```rust
-/// use horus::prelude::*;
-/// ```
-///
-/// This is the **only import** you need. All user-facing types, traits,
-/// macros, and message definitions are included.
-pub mod prelude {
-    // === Node ===
-    pub use horus_core::core::{HealthStatus, LogSummary, Node, NodeState};
-
-    // === Real-time node ===
-    pub use horus_core::core::rt_config::{
-        RtApplyResult, RtConfig, RtConfigBuilder, RtDegradation, RtKernelInfo, RtScheduler,
-    };
-    pub use horus_core::core::{DurationExt, Frequency, Miss, RtStats};
-
-    // === Rate / Stopwatch ===
-    pub use horus_core::core::timer::{Rate, Stopwatch};
-
-    // === Topic (IPC) ===
-    pub use horus_core::communication::Topic;
-
-    // === Scheduler ===
-    pub use horus_core::scheduling::Scheduler;
-
-    // === Execution configuration ===
-    pub use horus_core::scheduling::FailurePolicy;
-
-    // === Runtime parameters ===
-    pub use horus_core::params::RuntimeParams;
-
-    // === Coordinate transforms ===
-    // C++ gets `horus::TransformFrame` straight from <horus/horus.hpp> and
-    // horus_py exposes the same tree, because both crates depend on horus-tf.
-    // The Rust umbrella crate did not, so `horus::prelude` offered
-    // `TransformStamped` — the message — but not the frame tree that produces
-    // it, and a Rust user had to add a pinned git dependency by hand to do
-    // what the other two languages do out of the box.
-    pub use horus_tf::{FrameBuilder, Transform, TransformFrame, TransformFrameConfig};
-
-    // === Memory (domain types) ===
-    pub use horus_core::memory::{DepthImage, Image, PointCloud};
-
-    // === Universal IPC types (math, diagnostics, time, generic) ===
-    pub use horus_types::prelude::*;
-
-    // === Types (core) ===
-    /// Internal tensor dtype — prefer `PointCloud::from_xyz()`, `DepthImage::meters()` etc.
-    #[doc(hidden)]
-    pub use horus_core::types::TensorDtype;
-    pub use horus_core::types::{Device, ImageEncoding, PointXYZ, PointXYZI, PointXYZRGB};
-
-    // === Standard robotics messages ===
-    // Kept in the unified prelude so a normal `horus` application can use
-    // CmdVel, Imu, LaserScan, BatteryState, and related wire types without a
-    // second explicit dependency.
-    pub use horus_robotics::prelude::*;
-
-    // === Actions ===
-    pub use horus_core::actions::{
-        Action, ActionClient, ActionClientBuilder, ActionClientNode, ActionError, ActionResult,
-        ActionServerBuilder, ActionServerNode, CancelResponse, ClientGoalHandle, GoalId,
-        GoalOutcome, GoalPriority, GoalResponse, GoalStatus, PreemptionPolicy, ServerGoalHandle,
-    };
-
-    // === Services (request/response RPC) ===
-    pub use horus_core::services::{
-        AsyncServiceClient, Service, ServiceClient, ServiceError, ServiceRequest, ServiceResponse,
-        ServiceResult, ServiceServer, ServiceServerBuilder,
-    };
-
-    // === Errors ===
-    //
-    // `Error` and `Result` are short aliases (for use inside application code).
-    // `HorusError` is also exported so callers can pattern-match exhaustively:
-    //   `use horus::prelude::*;`
-    //   `if let Err(HorusError::InvalidDescriptor(msg)) = result { ... }`
-    pub use horus_core::error::{
-        retry_transient, CommunicationError, ConfigError, Error, HorusContext, HorusError,
-        MemoryError, NodeError, NotFoundError, ParseError, ResourceError, Result, RetryConfig,
-        SerializationError, Severity, TimeoutError, TransformError, ValidationError,
-    };
-
-    // === Time API ===
-    pub use crate::time::TimeStamp;
-
-    // === Macros ===
-    pub use horus_core::action;
-    pub use horus_core::hlog;
-    pub use horus_core::hlog_every;
-    pub use horus_core::hlog_once;
-    pub use horus_core::message;
-    pub use horus_core::service;
-    pub use horus_core::standard_action;
-    #[cfg(feature = "macros")]
-    pub use horus_macros::*;
-}
+pub mod prelude;
 
 // ─── Tests ───────────────────────────────────────────────────────────────────
